+++ core/src/main/java/org/seedstack/business/assembler/modelmapper/ModelMapperTupleAssembler.java
@@ -7,12 +7,9 @@
  */
 package org.seedstack.business.assembler.modelmapper;
 
-import net.jodah.typetools.TypeResolver;
 import org.javatuples.Tuple;
 import org.modelmapper.ModelMapper;
 import org.seedstack.business.assembler.AbstractBaseAssembler;
-import org.seedstack.business.assembler.BaseTupleAssembler;
-import org.seedstack.seed.core.utils.SeedReflectionUtils;
 
 /**
  * This assembler automatically assembles aggregates in DTO and vice versa.
@@ -22,22 +19,17 @@ import org.seedstack.seed.core.utils.SeedReflectionUtils;
  * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
  */
 public abstract class ModelMapperTupleAssembler<T extends Tuple, D> extends AbstractBaseAssembler<T, D> {
-
-    protected Class<D> dtoClass;
     private ModelMapper assembleModelMapper;
     private ModelMapper mergeModelMapper;
 
     @SuppressWarnings({"unchecked", "rawtypes"})
     public ModelMapperTupleAssembler() {
         initModelMappers();
-
-        Class<? extends BaseTupleAssembler> class1 = (Class<? extends BaseTupleAssembler>) SeedReflectionUtils.cleanProxy(getClass());
-        dtoClass = (Class<D>) TypeResolver.resolveRawArguments(class1.getGenericSuperclass(), class1)[1];
     }
 
     public ModelMapperTupleAssembler(Class<D> dtoClass) {
+        super(dtoClass);
         initModelMappers();
-        this.dtoClass = dtoClass;
     }
 
     @Override
@@ -68,7 +60,7 @@ public abstract class ModelMapperTupleAssembler<T extends Tuple, D> extends Abst
         }
     }
 
-    void initModelMappers() {
+    private void initModelMappers() {
         this.assembleModelMapper = new ModelMapper();
         configureAssembly(assembleModelMapper);